add color with hex
